cleaned up api/dataset route
