+++ src/app/page.tsx
@@ -10,12 +10,12 @@ export default function Home() {
         console.log('all datasets:', resDataset !== undefined)
         const resSingleDataset = await api.getDataset('Climate Change Impact')
         console.log('single dataset:', resSingleDataset !== undefined)
-        // const resEvent = await api.getEvent('wox79ra56il6u88') // swap out for a working id in your pb
-        // console.log('events:', resEvent)
         const resDatasetFromUserEvent = await api.getDatasetFromUserEvent(
             user?.id as string
         )
         console.log('datasetFromUser:', resDatasetFromUserEvent !== undefined)
+        const resEvent = await api.getEvent('wox79ra56il6u88') // swap out for a working id in your pb
+        console.log('events:', resEvent !== undefined)
     }
     run()
